requirements.txt
# Python dependencies for RAG-based Mutual Fund FAQ Chatbot
# Install: pip install -r requirements.txt
# Also run: python -m spacy download en_core_web_sm

# --- Google Gemini ---
google-generativeai>=0.7.0

# --- Vector Store ---
chromadb>=0.5.0

# --- HTTP & Scraping ---
httpx>=0.27.0
beautifulsoup4>=4.12.0
pdfminer.six>=20221105

# --- NLP ---
spacy>=3.7.0

# --- Retry & async ---
tenacity>=8.2.0

# --- Data ---
numpy>=1.26.0

# --- Environment ---
python-dotenv>=1.0.0

# --- UI ---
streamlit>=1.35.0

# --- Testing ---
pytest>=8.0.0
pytest-asyncio>=0.23.0

# --- Utilities ---
tqdm>=4.66.0
